Add error handling on main processing
This commit adds a couple of exceptions, and an overall exception handler that will make sure the process will exit with a non-zero status, indicating the actual process fails. This will make sure, if e.g. the spark-cli fails, or is missing, the process will stop in an errornous status that travis will understand.